requirements.txt
discord-webhook
requests>=2.31.0
pandas
schedule
pytz
python-dotenv>=1.0.0
requests>=2.31.0
psycopg2-binary>=2.9.9
numpy>=1.24.0
